REF : Rename test files
the test files are now named what unittest expects them to be, to be
discoverable.

a few imports have been updated where necessary
