+++ envisage/resource/tests/test_resource_manager.py
