+++ envisage/tests/test_class_load_hook.py
