+++ envisage/tests/test_composite_plugin_manager.py
