+++ envisage/tests/test_core_plugin.py
