+++ envisage/tests/test_egg_basket_plugin_manager.py
