+++ envisage/tests/test_extension_point_binding.py
