+++ envisage/tests/test_extension_registry.py
