+++ envisage/tests/test_import_manager.py
