+++ envisage/tests/test_package_plugin_manager.py
