+++ envisage/tests/test_plugin.py
