+++ envisage/tests/test_plugin_manager.py
