+++ envisage/tests/test_safeweakref.py
