+++ envisage/tests/test_slice.py
