+++ envisage/ui/action/tests/test_action_manager_builder.py
